requirements.txt
streamlit
langchain-community
ollama
faiss-cpu
langchain_ollama
sentence_transformers
pdfplumber
torch